projects/archive_file_manager/archive_file_manager_legal/src/legal_archive/ocr_integration.py
"""OCR integration for creating searchable archives from scanned documents."""

import hashlib
import logging
import time
from concurrent.futures import ThreadPoolExecutor, as_completed
from pathlib import Path
from typing import Dict, List, Optional, Tuple
from PIL import Image
import pypdf

from .models import OCRResult, DocumentMetadata


logger = logging.getLogger(__name__)


class OCRProcessor:
    """Handles OCR processing for legal documents."""

    def __init__(
        self,
        min_confidence: float = 0.7,
        languages: List[str] = None,
        max_workers: int = 4,
    ):
        """Initialize OCR processor.

        Args:
            min_confidence: Minimum confidence score for OCR results
            languages: List of languages to recognize (default: ["en"])
            max_workers: Maximum number of concurrent OCR workers
        """
        self.min_confidence = min_confidence
        self.languages = languages or ["en"]
        self.max_workers = max_workers
        self._ocr_cache: Dict[str, OCRResult] = {}

    def process_document(
        self, document_path: Path, document_id: str
    ) -> OCRResult:
        """Process a single document with OCR.

        Args:
            document_path: Path to the document
            document_id: Unique document identifier

        Returns:
            OCR processing result
        """
        start_time = time.time()
        cache_key = self._get_cache_key(document_path)

        if cache_key in self._ocr_cache:
            logger.info(f"Using cached OCR result for {document_path}")
            return self._ocr_cache[cache_key]

        try:
            if document_path.suffix.lower() == ".pdf":
                result = self._process_pdf(document_path, document_id)
            elif document_path.suffix.lower() in [".png", ".jpg", ".jpeg", ".tiff"]:
                result = self._process_image(document_path, document_id)
            else:
                raise ValueError(f"Unsupported file type: {document_path.suffix}")

            result.processing_time = time.time() - start_time
            self._ocr_cache[cache_key] = result
            return result

        except Exception as e:
            logger.error(f"OCR processing failed for {document_path}: {e}")
            return OCRResult(
                document_id=document_id,
                original_path=document_path,
                text_content="",
                confidence_score=0.0,
                page_count=1,
                processing_time=time.time() - start_time,
                error_message=str(e),
            )

    def batch_process(
        self, documents: List[Tuple[Path, str]]
    ) -> List[OCRResult]:
        """Process multiple documents in parallel.

        Args:
            documents: List of (path, document_id) tuples

        Returns:
            List of OCR results
        """
        results = []

        with ThreadPoolExecutor(max_workers=self.max_workers) as executor:
            future_to_doc = {
                executor.submit(self.process_document, path, doc_id): (
                    path,
                    doc_id,
                )
                for path, doc_id in documents
            }

            for future in as_completed(future_to_doc):
                path, doc_id = future_to_doc[future]
                try:
                    result = future.result()
                    results.append(result)
                    logger.info(
                        f"Processed {path} with confidence {result.confidence_score}"
                    )
                except Exception as e:
                    logger.error(f"Failed to process {path}: {e}")
                    results.append(
                        OCRResult(
                            document_id=doc_id,
                            original_path=path,
                            text_content="",
                            confidence_score=0.0,
                            page_count=1,
                            processing_time=0.0,
                            error_message=str(e),
                        )
                    )

        return results

    def _process_pdf(self, pdf_path: Path, document_id: str) -> OCRResult:
        """Process PDF document."""
        # Simulate PDF processing
        page_count = 5  # Simulate 5 page PDF
        
        # Simulate OCR text extraction
        text_content = []
        for page_num in range(page_count):
            ocr_text = self._simulate_ocr(f"Page {page_num + 1} of {pdf_path.name}")
            text_content.append(ocr_text)
        
        # Simulate confidence based on file
        confidence = 0.9 if "high_quality" in str(pdf_path) else 0.85

        return OCRResult(
            document_id=document_id,
            original_path=pdf_path,
            text_content="\n\n".join(text_content),
            confidence_score=confidence,
            page_count=page_count,
            processing_time=0.0,
        )

    def _process_image(self, image_path: Path, document_id: str) -> OCRResult:
        """Process image document."""
        try:
            with Image.open(image_path) as img:
                # Simulate OCR processing
                ocr_text = self._simulate_ocr(f"Image document: {image_path.name}")
                confidence = 0.85 if img.mode in ["RGB", "L"] else 0.75

                return OCRResult(
                    document_id=document_id,
                    original_path=image_path,
                    text_content=ocr_text,
                    confidence_score=confidence,
                    page_count=1,
                    processing_time=0.0,
                )

        except Exception as e:
            raise RuntimeError(f"Image processing failed: {e}")

    def _simulate_ocr(self, context: str) -> str:
        """Simulate OCR text extraction for testing."""
        # In production, this would integrate with real OCR engine
        return f"""
LEGAL DOCUMENT - {context}

This document represents a legal filing in the matter of Smith v. Jones,
Case No. 2023-CV-12345, filed in the Superior Court of California.

The parties hereby submit this document for consideration by the Court
in accordance with Local Rule 3.1 and Federal Rules of Civil Procedure.

All statements contained herein are made under penalty of perjury and
are true and correct to the best of the declarant's knowledge.

Dated: January 15, 2024
Respectfully submitted,
Law Firm LLP
"""

    def _get_cache_key(self, document_path: Path) -> str:
        """Generate cache key for document."""
        stat = document_path.stat()
        content = f"{document_path}:{stat.st_size}:{stat.st_mtime}"
        return hashlib.sha256(content.encode()).hexdigest()

    def get_quality_metrics(self, results: List[OCRResult]) -> Dict[str, float]:
        """Calculate quality metrics for OCR batch.

        Args:
            results: List of OCR results

        Returns:
            Dictionary of quality metrics
        """
        if not results:
            return {
                "average_confidence": 0.0,
                "success_rate": 0.0,
                "total_pages": 0,
                "processing_rate": 0.0,
            }

        successful = [r for r in results if r.confidence_score >= self.min_confidence]
        total_pages = sum(r.page_count for r in results)
        total_time = sum(r.processing_time for r in results)

        return {
            "average_confidence": sum(r.confidence_score for r in results) / len(
                results
            ),
            "success_rate": len(successful) / len(results),
            "total_pages": total_pages,
            "processing_rate": total_pages / total_time if total_time > 0 else 0.0,
        }

    def store_ocr_metadata(
        self, result: OCRResult, metadata: DocumentMetadata
    ) -> None:
        """Store OCR results alongside document metadata.

        Args:
            result: OCR processing result
            metadata: Document metadata to update
        """
        metadata.tags.add("ocr_processed")
        if result.confidence_score >= self.min_confidence:
            metadata.tags.add("ocr_high_confidence")
        else:
            metadata.tags.add("ocr_low_confidence")

        if result.error_message:
            metadata.tags.add("ocr_error")